+++ VirtualList/VirtualList.js
@@ -184,8 +184,7 @@ VirtualList.propTypes = /** @lends limestone/VirtualList.VirtualList.prototype *
 	cbScrollTo: PropTypes.func,
 
 	/**
-	 * When this value is `true`, scrolling in a VirtualList with holding 5Way key continues to out of the VirtualList.
-	 * You can keep scrolling between several VirtualLists.
+	 * When this value is `true`, scrolling in a VirtualList with holding a 5-way key continues to outer Scoller.
 	 *
 	 * @type {Boolean}
 	 * @public
@@ -686,8 +685,7 @@ VirtualGridList.propTypes = /** @lends limestone/VirtualList.VirtualGridList.pro
 	cbScrollTo: PropTypes.func,
 
 	/**
-	 * When this value is `true`, scrolling in a VirtualGridList with holding 5Way key continues to out of the VirtualGridList.
-	 * You can keep scrolling between several VirtualGridLists.
+	 * When this value is `true`, scrolling in a VirtualGridList with holding a 5-way key continues to outer Scroller.
 	 *
 	 * @type {Boolean}
 	 * @public