+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 ### Added
 
-- `limestone/VirtualList` prop `continue5WayHold` to scroll continuously between VirtualLists
+- `limestone/VirtualList` prop `continue5WayHold` to scroll continuously from in a VirtualList to the outer scroller.
 
 ### Fixed